+++ Source/Scene/ModelUtility.js
@@ -473,8 +473,9 @@ define([
         'KHR_draco_mesh_compression' : true,
         'KHR_materials_common' : true,
         'KHR_techniques_webgl' : true,
-        'WEB3D_quantized_attributes' : true,
-        'KHR_materials_unlit' : true
+        'KHR_materials_unlit' : true,
+        'KHR_materials_pbrSpecularGlossiness' : true,
+        'WEB3D_quantized_attributes' : true
     };
 
     ModelUtility.checkSupportedExtensions = function(extensionsRequired) {