+++ Specs/Scene/ModelSpec.js
@@ -119,6 +119,9 @@ defineSuite([
     var CesiumManUrl = './Data/Models/MaterialsCommon/Cesium_Man.gltf';
 
     var boomBoxUrl = './Data/Models/PBR/BoomBox/BoomBox.gltf';
+    var boomBoxPbrSpecularGlossinessUrl = './Data/Models/PBR/BoomBoxSpecularGlossiness/BoomBox.gltf';
+    var boomBoxPbrSpecularGlossinessDefaultsUrl = './Data/Models/PBR/BoomBoxSpecularGlossiness/BoomBox-Default.gltf';
+    var boomBoxPbrSpecularGlossinessNoTextureUrl = './Data/Models/PBR/BoomBoxSpecularGlossiness/BoomBox-NoTexture.gltf';
     var boxPbrUrl = './Data/Models/PBR/Box/Box.gltf';
     var boxPbrUnlitUrl = './Data/Models/PBR/BoxUnlit/BoxUnlit.gltf';
     var boxAnimatedPbrUrl = './Data/Models/PBR/BoxAnimated/BoxAnimated.gltf';
@@ -2715,7 +2718,7 @@ defineSuite([
         });
     });
 
-    it('renders with the unlit extension', function() {
+    it('renders with the KHR_materials_unlit extension', function() {
         return loadModel(boxPbrUnlitUrl).then(function(model) {
             model.show = true;
             model.zoomTo();
@@ -2727,6 +2730,33 @@ defineSuite([
             });
 
             primitives.remove(model);
+        });
+    });
+
+    it('renders with the KHR_materials_pbrSpecularGlossiness extension', function() {
+        return loadModel(boomBoxPbrSpecularGlossinessUrl).then(function(model) {
+            model.show = true;
+            model.zoomTo();
+
+            scene.renderForSpecs();
+        });
+    });
+
+    it('renders with the KHR_materials_pbrSpecularGlossiness extension with defaults', function() {
+        return loadModel(boomBoxPbrSpecularGlossinessDefaultsUrl).then(function(model) {
+            model.show = true;
+            model.zoomTo();
+
+            scene.renderForSpecs();
+        });
+    });
+
+    it('renders with the KHR_materials_pbrSpecularGlossiness extension when no textures are found', function() {
+        return loadModel(boomBoxPbrSpecularGlossinessNoTextureUrl).then(function(model) {
+            model.show = true;
+            model.zoomTo();
+
+            scene.renderForSpecs();
         });
     });